Extract sorted-serving eligibility out of ReplayServingFactory
Move the stamp/mode/completeness/index-derive decision into its own
SortedEligibility class (io.varve.swath.replay.fixture) so a second caller
outside :swath-replay-server's server package can reuse the exact decision
instead of restating it. ReplayServingFactory now delegates to it; behavior
is unchanged.
